add bubble hook
